In this PR, I:
- Added `LoadingUnitDetails.AreaFillRate` that indicates the percentage of occupied area
- Added Controller endpoint to get loading units by aisle `AisleController.GetLoadingUnitsById()`
- Added related provider method `LoadingUnitsProvider.GetByAisleId()`

Also, I performed these refactorings:
- Changed data type of warehouse quantities from `int` to `double`
- Removed `ItemListItemsCount` from List Row BOs because of promiscuous measure units

Related work items: #2269

+++ Common/DataModels/Mission.cs
@@ -21,7 +21,7 @@ namespace Ferretto.Common.DataModels
 
         public DateTime CreationDate { get; set; }
 
-        public int DispatchedQuantity { get; set; }
+        public double DispatchedQuantity { get; set; }
 
         public int Id { get; set; }
 
@@ -57,7 +57,7 @@ namespace Ferretto.Common.DataModels
 
         public string RegistrationNumber { get; set; }
 
-        public int RequestedQuantity { get; set; }
+        public double RequestedQuantity { get; set; }
 
         public MissionStatus Status { get; set; }
