+++ WMS/Data/Core/Models/SchedulerRequest/SchedulerRequest.cs
@@ -39,9 +39,9 @@ namespace Ferretto.WMS.Data.Core.Models
 
         public string RegistrationNumber { get; set; }
 
-        public int? RequestedQuantity { get; set; }
+        public double? RequestedQuantity { get; set; }
 
-        public int? ReservedQuantity { get; set; }
+        public double? ReservedQuantity { get; set; }
 
         public SchedulerRequestStatus Status { get; set; }
